backend/requirements.txt
# FastAPI & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.4

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# AI & ML (API-based)
openai>=1.50.0
tiktoken==0.5.2

# Background Tasks
celery==5.3.6
redis==5.0.1
apscheduler==3.10.4

# Communication
sendgrid==6.11.0
twilio==8.10.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
httpx==0.26.0
python-dotenv==1.0.0
structlog==24.1.0
tenacity==8.2.3

# PDF / Document Processing
pypdf==3.17.4
python-docx==1.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
